requirements.txt
pyserial
flask
flask-cors